requirements.txt
argcomplete==1.10.3
beautifulsoup4==4.8.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
compressed-rtf==1.0.6
docx==0.2.4
docx2txt==0.8
easyocr==1.6.2
ebcdic==1.1.1
et-xmlfile==1.1.0
extract-msg==0.28.7
idna==3.4
imageio==2.24.0
IMAPClient==2.1.0
lxml==4.9.2
networkx==3.0
ninja==1.11.1
numpy==1.24.1
olefile==0.46
opencv-contrib-python==4.5.5.62
opencv-python-headless==4.5.4.60
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
pdfminer.six==20191110
Pillow==9.4.0
pyclipper==1.3.0.post4
pycryptodome==3.16.0
pyocr==0.8.3
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.0
shapely==2.0.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
textract==1.6.5
tifffile==2022.10.10
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
xlrd==1.2.0
XlsxWriter==3.0.7
xlwt==1.3.0
